+++ app/login/page.tsx
@@ -55,7 +55,7 @@ export default function LoginPage() {
       </div>
 
       {/* Auth UI */}
-      <div className="relative z-10 pt-28">
+      <div className="relative z-10 pt-20 sm:pt-28">
 
         <GuildAuthUI />